pyproject.toml
[tool.poetry]
name = "ZenBot"
version = "0.1.0"
description = ""
authors = ["Daniel Munch <daniellmunch@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.3"
DavesLogger = "^1.1.0"
asyncio = "^3.4.3"
aiohttp = "^3.6.2"
requests = "^2.23.0"
motor = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[tool.poetry.scripts]
start = "zenbot.main:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
